dev (#1)
XMMS now in working state, can play music and doesn't crash every other second

Reviewed-on: https://git.yeold.org/yeold/xmms-revived/pulls/1

+++ Input/wav/wav.h
@@ -47,10 +47,9 @@ extern InputPlugin wav_ip;
 typedef struct
 {
 	FILE *file;
-	short format_tag, channels, block_align, bits_per_sample, eof,
-	      going;
+	short format_tag, channels, block_align, bits_per_sample, eof, going;
 	long samples_per_sec, avg_bytes_per_sec;
-	int position, length;
+	unsigned long position, length;
 	int seek_to, data_offset;
 	pid_t pid;
 }
